apis/components/requirements.txt
flask
flask_restful
confluent_kafka
psycopg2-binary
hvac
